+++ .github/workflows/ci.yaml
@@ -162,7 +162,7 @@ jobs:
           mvn -B -q install --file pom.xml -P dse -Dit.test=${{ matrix.itTest }} -DfailIfNoTests=false \
           -DrunDSEtests=${{ matrix.runDSEtests }} \
           -DrunDSEtestsUBI=${{ matrix.runDSEtestsUBI }}
-  build-dse7:
+  build-hcd:
     if: github.repository_owner == 'k8ssandra'
     runs-on: ubuntu-latest
     steps:
@@ -209,7 +209,7 @@ jobs:
           </settings>
           EOF
           cp ~/.m2/settings.xml settings.xml
-          mvn -B -q install --file pom.xml -P dse7
+          mvn -B -q install --file pom.xml -P hcd
 
   publish-oss:
     name: Publish ${{ matrix.cassandra-version }} Cassandra image
