+++ .github/workflows/jar-release.yaml
@@ -40,7 +40,7 @@ jobs:
            </servers>
           </settings>
           EOF
-          mvn -q -ff package -DskipTests -P dse,dse7
+          mvn -q -ff package -DskipTests -P dse,hcd
       - name: zip-up
         run: |
           zip jars.zip \
@@ -51,7 +51,7 @@ jobs:
           management-api-agent-4.1.x/target/datastax-mgmtapi-agent-*.jar \
           management-api-agent-5.0.x/target/datastax-mgmtapi-agent-*.jar \
           management-api-agent-dse-6.8/target/datastax-mgmtapi-agent-*.jar \
-          management-api-agent-dse7/target/datastax-mgmtapi-agent-*.jar \
+          management-api-agent-hcd/target/datastax-mgmtapi-agent-*.jar \
           management-api-server/target/datastax-mgmtapi-server-*.jar \
           management-api-common/target/datastax-mgmtapi-common-*.jar
       - name: Release
