--- management-api-agent-dse7/src/main/java/com/datastax/mgmtapi/CassandraAPIServiceProviderDse7.java
@@ -1,17 +1,0 @@
-/*
- * Copyright DataStax, Inc.
- *
- * Please see the included license file for details.
- */
-package com.datastax.mgmtapi;
-
-import com.datastax.mgmtapi.shim.CassandraAPIDse7;
-import com.datastax.mgmtapi.shims.CassandraAPI;
-
-public class CassandraAPIServiceProviderDse7 implements CassandraAPIServiceProvider {
-
-  @Override
-  public CassandraAPI getCassandraAPI() {
-    return new CassandraAPIDse7();
-  }
-}
