--- management-api-agent-dse7/src/main/java/com/datastax/mgmtapi/rpc/RpcMethodServiceProviderDse7.java
@@ -1,16 +1,0 @@
-/*
- * Copyright DataStax, Inc.
- *
- * Please see the included license file for details.
- */
-package com.datastax.mgmtapi.rpc;
-
-import java.lang.reflect.Method;
-
-public class RpcMethodServiceProviderDse7 implements RpcMethodServiceProvider {
-
-  @Override
-  public RpcMethod getRpcMethod(Method method, RpcObject rpcObject) {
-    return new RpcMethodDse7(method, rpcObject);
-  }
-}
