--- management-api-agent-dse7/src/main/resources/META-INF/services/com.datastax.mgmtapi.CassandraAPIServiceProvider
@@ -1,1 +1,0 @@
-com.datastax.mgmtapi.CassandraAPIServiceProviderDse7
