+++ management-api-agent-hcd/src/main/java/com/datastax/mgmtapi/CassandraAPIServiceProviderHcd.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright DataStax, Inc.
+ *
+ * Please see the included license file for details.
+ */
+package com.datastax.mgmtapi;
+
+import com.datastax.mgmtapi.shim.CassandraAPIHcd;
+import com.datastax.mgmtapi.shims.CassandraAPI;
+
+public class CassandraAPIServiceProviderHcd implements CassandraAPIServiceProvider {
+
+  @Override
+  public CassandraAPI getCassandraAPI() {
+    return new CassandraAPIHcd();
+  }
+}
