+++ management-api-agent-hcd/src/main/java/com/datastax/mgmtapi/rpc/RpcMethodServiceProviderHcd.java
@@ -1,0 +1,16 @@
+/*
+ * Copyright DataStax, Inc.
+ *
+ * Please see the included license file for details.
+ */
+package com.datastax.mgmtapi.rpc;
+
+import java.lang.reflect.Method;
+
+public class RpcMethodServiceProviderHcd implements RpcMethodServiceProvider {
+
+  @Override
+  public RpcMethod getRpcMethod(Method method, RpcObject rpcObject) {
+    return new RpcMethodHcd(method, rpcObject);
+  }
+}
