+++ management-api-agent-hcd/src/main/resources/META-INF/services/com.datastax.mgmtapi.CassandraAPIServiceProvider
@@ -1,0 +1,1 @@
+com.datastax.mgmtapi.CassandraAPIServiceProviderHcd
