+++ management-api-agent-hcd/src/main/resources/META-INF/services/com.datastax.mgmtapi.rpc.RpcMethodServiceProvider
@@ -1,0 +1,1 @@
+com.datastax.mgmtapi.rpc.RpcMethodServiceProviderHcd
